requirements.txt
coop-geo==0.2.0
Django==1.11.9
django-autoslug==1.7.2
django-bitfield==1.8.0
django-extensions==1.9.9
django-floppyforms==0.4.7
django-geojson==2.11.0
django-jsonfield==0.9.15
django-leaflet==0.23.0
django-model-utils==2.2
django-mptt==0.7.4
django-separatedvaluesfield==0.3.2
django-taggit==0.22.2
django-timezone-field==1.3
django-uuidfield==0.5.0
djangorestframework==3.7.7
gunicorn==19.3.0
numpy==1.10.1
psycopg2==2.6.1
pytz==2015.6
raven==5.8.1
requests==2.8.1
simplejson==3.6.5
six==1.10.0
wheel==0.24.0
django-taggit-serializer==0.1.5
django-autocomplete-light==3.2.10
django-cors-headers==2.1.0
nltk==3.0.3
nltk==3.0.3
slugify==0.0.1
celery==4.1.0
https://github.com/aljosa/django-tinymce/archive/master.zip
django-filter==1.0.4
django-filters==0.2.1
celery==4.1.0
django-tinymce==2.6.0
Lemmagen==1.2.0
editdistance==0.4
coreapi==2.3.3